Move PUBLICATION_URI into app/standard-site.js
Every consumer now reaches the module anyway for the document helpers, so
the route was an odd place for the publication URI to live — importers of
one constant were pulling in a Next.js route handler to get it.

The DID becomes the single literal and both AT-URIs derive from it, which
also drops the split('/') that recovered it from the publication URI.
