+++ app/.well-known/site.standard.publication/route.js
@@ -1,16 +1,8 @@
-// standard.site (https://standard.site) publishes AT Protocol lexicons for
-// long-form writing. A `site.standard.publication` record describes this blog,
-// and is only trusted once this endpoint points back at it. Verifiers fetch
-// this and compare the body to the record's AT-URI.
-// See https://standard.site/docs/verification.
-//
-// did:plc:mnbueka7mnf5ts5rzmjy34z2 is the DID behind the @philihp.com handle.
-// Prefer the DID over the handle here — handles can be reassigned.
-export const PUBLICATION_URI =
-  'at://did:plc:mnbueka7mnf5ts5rzmjy34z2/site.standard.publication/self'
+import { PUBLICATION_URI } from '../../standard-site.js'
 
 export const dynamic = 'force-static'
 
+// Verifiers fetch this and compare the body to the publication record's AT-URI.
 export const GET = () =>
   new Response(`${PUBLICATION_URI}\n`, {
     headers: {
