+++ app/standard-site.js
@@ -1,7 +1,15 @@
-import { PUBLICATION_URI } from './.well-known/site.standard.publication/route.js'
+// standard.site (https://standard.site) publishes AT Protocol lexicons for
+// long-form writing: a `site.standard.publication` record describing this blog,
+// and a `site.standard.document` record per post.
+//
+// did:plc:mnbueka7mnf5ts5rzmjy34z2 is the DID behind the @philihp.com handle.
+// Prefer the DID over the handle — handles can be reassigned.
+export const DID = 'did:plc:mnbueka7mnf5ts5rzmjy34z2'
 
-// at://<did>/site.standard.publication/self
-const DID = PUBLICATION_URI.split('/')[2]
+// Served verbatim from /.well-known/site.standard.publication, which is what
+// ties the record to this domain, and echoed as a discovery hint in <head>.
+// See https://standard.site/docs/verification.
+export const PUBLICATION_URI = `at://${DID}/site.standard.publication/self`
 
 // Record keys cannot contain slashes, so a post's route doubles as its rkey
 // with the separators swapped: /2026/announcing-pointille.html becomes
